fix: visualize marker in default
